updated main.js and data.js
