+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomInteger, getUniqNumberFromRange} from './util.js';
+
 const URL = [];
 const PHOTOS_ID = [];
 const COMMENTS_ID = [];
@@ -31,14 +33,30 @@ const DESCRIPTIONS = [
   'От твоего взгляда моё сердце дрожит, как пустой холодильник.'
 ];
 
-export {URL,
-  PHOTOS_ID,
-  COMMENTS_ID,
-  NUMBER_OF_POSTS,
-  MINIMUM_NUMBER_OF_COMMENTS,
-  MAXIMUM_NUMBER_OF_COMMENTS,
-  MINIMUM_NUMBER_OF_LIKES,
-  MAXIMUM_NUMBER_OF_LIKES,
-  TEXT_MESSAGE,
-  NAMES,
-  DESCRIPTIONS};
+const getComments = () => {
+  const comments = [];
+  const numberOfComments = getRandomInteger(MINIMUM_NUMBER_OF_COMMENTS, MAXIMUM_NUMBER_OF_COMMENTS);
+  for (let i=0; i < numberOfComments; i++){
+    comments.push({
+      id: getUniqNumberFromRange(1, 100000, COMMENTS_ID),
+      avatar:`img/avatar-${getRandomInteger(1, 6)}.svg`,
+      message: TEXT_MESSAGE[getRandomInteger(0, 5)],
+      name: NAMES[getRandomInteger(0, 5)]
+    });
+  }
+  return comments;
+};
+
+const getPost= () => ({
+  id: getUniqNumberFromRange(1, 25, PHOTOS_ID),
+  url: `photos/${getUniqNumberFromRange(1, 25, URL)}.jpg`,
+  description: DESCRIPTIONS[getRandomInteger(0, 4)],
+  likes: getRandomInteger(MINIMUM_NUMBER_OF_LIKES, MAXIMUM_NUMBER_OF_LIKES),
+  comments: getComments()
+});
+
+const getPosts = () => {
+  Array.from({length: NUMBER_OF_POSTS}, getPost);
+};
+
+export {getPosts};
